* Update to geos-sys 2.0 * Strenghten checks on *const/*mut usage by creating a new AsRawMut trait * Create a new ConstGeometry type to better handle the split with Geometry

+++ src/enums.rs
@@ -301,8 +301,8 @@ impl TryFrom<size_t> for Ordinate {
     }
 }
 
-impl Into<size_t> for Ordinate {
-    fn into(self) -> size_t {
+impl Into<u32> for Ordinate {
+    fn into(self) -> u32 {
         match self {
             Ordinate::X => 0,
             Ordinate::Y => 1,
